additional merkle ops
